python/requirements.txt
google-generativeai>=0.3.0
PyMuPDF>=1.23.0
Pillow>=10.0.0

